requirements.txt
tensorflow-cpu==2.13.0
streamlit==1.32.2
pillow
numpy==1.23.5
